Rename files to maintain continuity
