--- wsg_50_simulation/controllers/wsg50_driver.yaml
@@ -1,41 +1,0 @@
-controller_manager:
-  ros__parameters:
-    update_rate: 500  # Hz
-
-    joint_state_broadcaster:
-      type: joint_state_broadcaster/JointStateBroadcaster
-
-    wsg_50_gl:
-      type: position_controllers/JointGroupPositionController
-    
-    wsg_50_gr:
-      type: position_controllers/JointGroupPositionController
-    
-    position_controller:
-      type: position_controllers/JointGroupPositionController
-
-wsg_50_gl:
-  ros__parameters:
-    joints: 
-      - wsg50_base_joint_gripper_left
-    pid:
-      p: 100.0
-      d: 10.0
-
-wsg_50_gr:
-  ros__parameters:
-    joints: 
-      - wsg50_base_joint_gripper_right
-    pid:
-      p: 100.0
-      d: 10.0
-
-position_controller:
-  ros__parameters:
-    joints: 
-      - wsg50_base_joint_gripper_right
-      - wsg50_base_joint_gripper_left
-    pid:
-      p: 100.0
-      d: 10.0
-
