+++ wsg_50_simulation/controllers/wsg50_integrate.yaml
@@ -1,7 +1,6 @@
 left_wsg_50_gl:
   ros__parameters:
     type: position_controllers/JointGroupPositionController
-
     joints: 
       - left_gripper_base_joint_gripper_left
     pid:
